backend/requirements.txt
fastapi
uvicorn
ddgs
google-generativeai
python-dotenv
requests
beautifulsoup4
